requirements.txt
asgiref==3.4.1
bitrix24-rest==1.1.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==3.2.8
django-recaptcha==2.0.6
idna==3.3
python-decouple==3.5
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
psycopg2==2.9.5
django-cleanup==5.1.0
django-uuslug==2.0.0
Pillow==9.3.0
mysqlclient==2.1.1
djangorestframework==3.14.0
markdown==3.4.1
django-filter==22.1